# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TeamUp API
  description: |
    Sports venue booking and team formation API

    Features:
    - Venue search with geolocation
    - Timeslot booking
    - TeamUp creation and management
    - Join request workflow
  version: 1.0.0
  contact:
    name: TeamUp API Support
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.teamup.example.com
    description: Production server
...
paths:
...
  /auth/signup:
    post:
      tags: [Auth]
      summary: Sign up new user
      operationId: signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupIn'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenOut'
        '409':
          description: Email already exists
  /auth/login:
    post:
      tags: [Auth]
      summary: Login user
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginIn'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenOut'
        '401':
          description: Invalid credentials
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user info
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
        '401':
          description: Unauthorized
  # ============================================================================
  # VENUE ROUTES
  # ============================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login/signup
  schemas:
...
    SignupIn:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          minLength: 6
          example: password123
        display_name:
          type: string
          nullable: true
          example: John Doe
    LoginIn:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    TokenOut:
      type: object
      required: [access_token, refresh_token]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
    UserOut:
      type: object
      required: [id, email, display_name, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        display_name:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # ========================================================================
    # VENUE SCHEMAS
    # ========================================================================
...